appSlice refactor: rename folder from AppAlert to appAlert and fix import
