+++ features/index.ts
@@ -1,6 +1,6 @@
 export * from './ConfirmedEmail'
-export * from './AppAlert'
 export * from './ExpiredEmailLink'
 export * from './ForgotPassword'
+export * from './appAlert'
 export * from './newPassword'
 export * from './signUp'
